% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{datapackage-package}
\alias{datapackage}
\alias{datapackage-package}
\title{datapackage: Store Test Data for Quarto-WebR Usage}
\description{
Your package description. It must end with a period (".") and include relevant bibliographical references if applicable, using the following format: Author et al. (2023) \doi{10.5281/zenodo.6619350}.
}
\author{
\strong{Maintainer}: Karim Mané \email{karim.mane@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-9892-2999}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item London School of Hygiene and Tropical Medicine, LSHTM (00a0jsq62) [copyright holder]
}

}
\keyword{internal}
